bug fixed

See merge request artemnadosha/stepproject-cards!28

+++ src/js/tools/fetch/fetchDeleteCard.js
@@ -1,8 +1,12 @@
-import { request } from "./fetch.js";
+import { request } from './fetch.js';
 
-export const fetchDeleteCard = async (id) => {
-  return await request({
-    url: `/${id}`,
-    method: "DELETE",
-  });
+export const fetchDeleteCard = async id => {
+    if (!id) {
+        alert('Виникла помилка перезавантажте сторінку');
+        return;
+    }
+    return await request({
+        url: `/${id}`,
+        method: 'DELETE',
+    });
 };
